@@ -8,8 +8,8 @@
     "prebuild": "run-s clean eslint",
     "build": "tsc",
     "eslint": "eslint --fix \"src/**/*.ts\"",
-    "start": "node build/index.js",
-    "test": "mocha \"src/**/*.unit.test.ts\"",
+    "start": "node build/src/index.js",
+    "test": "ts-mocha \"src/**/*.unit.test.ts\"",
     "check": "npm run eslint",
     "coverage": "nyc npm test",
     "precoverage": "npm run check",
@@ -20,7 +20,7 @@
     "url": "git+https://github.com/danosaure/photo-timeline.git"
   },
   "engines": {
-    "node": ">=16.16.0"
+    "node": ">=18.14.0"
   },
   "keywords": [
     "photo",
@@ -37,17 +37,18 @@
   },
   "homepage": "https://github.com/danosaure/photo-timeline#readme",
   "dependencies": {
-    "argparse": "^2.0.1",
+    "@types/yargs": "^17.0.22",
     "babel-regenerator-runtime": "^6.5.0",
     "bluebird": "~3.7.2",
     "debug": "^4.3.4",
     "dist-exiftool": "^10.53.0",
     "exif": "^0.6.0",
-    "fast-exif": "~1.0.1",
+    "exifreader": "^4.9.1",
     "file-type": "^18.2.0",
     "lokijs": "~1.5.9",
     "md5-file": "~5.0.0",
-    "read-chunk": "^4.0.3"
+    "read-chunk": "^4.0.3",
+    "yargs": "^17.6.2"
   },
   "devDependencies": {
     "@babel/cli": "^7.20.7",
@@ -65,13 +66,13 @@
     "chai": "^4.3.7",
     "dirty-chai": "^2.0.1",
     "eslint": "^8.33.0",
-    "eslint-config-airbnb-base": "^15.0.0",
     "eslint-plugin-import": "^2.27.5",
     "mocha": "^10.2.0",
     "npm-run-all": "^4.1.5",
     "nyc": "^15.1.0",
     "rimraf": "^4.1.2",
     "source-map-support": "^0.5.19",
+    "ts-mocha": "^10.0.0",
     "typescript": "^4.9.5"
   },
   "nyc": {
@@ -80,11 +81,11 @@
     "report-dir": "./reports/coverage",
     "check-coverage": false,
     "include": [
-      "lib/**/*.js",
-      "src/**/*.js"
+      "lib/**/*.ts",
+      "src/**/*.ts"
     ],
     "exclude": [
-      "**/*.*.test.js"
+      "**/*.*.test.ts"
     ],
     "reporter": [
       "cobertura",
