@@ -1,44 +1,34 @@
-import { ArgumentParser } from 'argparse';
+import yargs from 'yargs';
+import { hideBin } from 'yargs/helpers';
 
 import { DEFAULT_LIMIT, EXTENSION } from './constants';
-import packageJson from '../package.json';
 
-const parser = new ArgumentParser({
-  add_help: true,
-  description: 'Sort folder of media files and rename them to date structure.',
-});
-
-parser.add_argument('-v', '--version', {
-  action: 'version',
-  version: packageJson.version,
-});
-
-parser.add_argument('-d', '--debug', {
-  help: 'Enable debug mode',
-  action: 'store_true',
-  default: false,
-});
-
-parser.add_argument('-l', '--limit', {
-  help: `Limit the number of files to process. -1 for all of them. (Default: ${DEFAULT_LIMIT})`,
-  default: DEFAULT_LIMIT,
-  type: 'int',
-});
-
-parser.add_argument('-e', '--ext', {
-  help: 'List of extensions to process',
-  required: true,
-  action: 'append',
-  choices: Object.values(EXTENSION).sort(),
-});
-
-parser.add_argument('-t', '--target', {
-  help: 'Target folder.',
-  required: true,
-});
-
-parser.add_argument('source', {
-  nargs: '+',
-});
-
-export default parser;
+export default async () => yargs(hideBin(process.argv))
+  .options({
+    limit: {
+      alias: 'l',
+      type: 'number',
+      description: `Limit the number of files to process. -1 for all of them. (Default: ${DEFAULT_LIMIT})`,
+      default: DEFAULT_LIMIT,
+    },
+    ext: {
+      alias: 'e',
+      type: 'string',
+      choices: Object.values(EXTENSION).sort(),
+      demandOption: true,
+      description: 'List of extensions to process',
+    },
+    source: {
+      alias: 's',
+      type: 'string',
+      description: 'Source folder',
+      demandOption: true,
+    },
+    target: {
+      alias: 't',
+      type: 'string',
+      description: 'Target folder.',
+      demandOption: true,
+    },
+  })
+  .argv;
