@@ -1,28 +1,20 @@
-// import cr2 from './cr2';
 import jpg from './jpg';
-// import mov from './mov';
 
 import { EXTENSION, EXTENSIONS } from '../constants';
 import fileExtention from '../file-extension';
 
-import _debug from './debug';
-
-const debug = _debug(__filename);
-
 export default async (filePath:string, fileTypes:string) => {
-  debug(`(filePath=${filePath}, fileTypes=${fileTypes})`);
+  console.log(`process-file/index: (filePath="${filePath}", fileTypes="${fileTypes}")`);
   const ext = fileExtention(filePath);
-  debug('ext=', ext);
 
+  let info = null;
   if (fileTypes.includes(EXTENSION.JPG) && EXTENSIONS.JPG.includes(ext)) {
-    return jpg(filePath, EXTENSION.JPG);
+    try {
+      info = await jpg(filePath, EXTENSION.JPG);
+    } catch {
+      console.log(`process-file/index:     need to quarantine filePath="${filePath}".`)
+    }
   }
-
-  return null;
+  console.log(`process-file/index: info=`, info);
+  return info;
 };
-
-// module.exports = Object.freeze({
-//     cr2: require('./cr2'),
-//     jpg: require('./jpg'),
-//     mov: require('./mov'),
-// });
